Forgetting rate proportional to average priority of the Bag. Modified: return type of forgetRate() and added second constructor to ConceptBag.java to allow average priority participate in forgetting rate

+++ nars_core_java/nars/storage/TaskLinkBag.java
@@ -52,7 +52,7 @@ public class TaskLinkBag extends Bag<TaskLink> {
      * @return The forget rate of TaskLinkBag
      */
     @Override
-    protected int forgetRate() {
+    protected float forgetRate() {
         return memory.getTaskForgettingRate().get();
     }
 }
